deploy/requirements.txt
azure-batch==9.0.0
azure-storage-file-share==12.3.0
